+++ RELEASE-NOTES.md
@@ -46,6 +46,7 @@ Project Page: https://github.com/a2stuff/a2d
 * Fix behavior after File > Duplicate failed due to space or storage type.
 * Set empty window maprect so that when restored with new file scrollbars are inactive.
 * Fix volume icon flicker after Check All Drives for removable disks.
+* Select volume icomn after Format/Erase.
 
 ### Selector